get post de mascotas/Validation

+++ src/main/java/ph/edu/cibertec/veterinaria/repositorios/MascotaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ph.edu.cibertec.veterinaria.entidades.Mascota;
 
-public interface MascotaRepository extends JpaRepository<Mascota, Integer> {
+public interface MascotaRepository extends JpaRepository<Mascota, Long> {
 
 }
